refactor: split TimelineRuler and TimelineSlider into <150-line units (#196)
TimeLine.tsx (TimelineRuler, ~1430 lines) -> app/components/timeline-ruler/:
- Hooks: usePlayhead, useTimelineZoom, useTrimSelection, useTimelineHistory,
  useTimelineProgress, scissor/segment/zoom/text/handle drag hooks, trim/zoom
  action hooks, and a useTimelineRuler orchestrator returning a view-model.
- Components: TimelineToolbar, TimelineTrack (+ SliceHandle/RulerLayers),
  TimelineBlocks (trim/zoom/text), TimelineRulerView.

MytimeLine.tsx (TimelineSlider, ~933 lines) -> app/components/mytimeline/:
- Hooks: useSvgWidth, useTimelineSliderDrag, useSegmentReorder,
  useTimelineKeyboard, useSegmentsCore (+ sync effects), geometry hook, and a
  useTimelineSlider orchestrator.
- Components: SliderToolbar, SegmentButtons, TimelineSvg (+ ZoomEffectMarkers,
  TimelineTicks, TimelineHandles), TimelineSliderView; shared constants/helpers.

Default/named exports preserved. Stable refs/setters added to dependency arrays
where they became hook params to avoid new exhaustive-deps warnings.

Verified: npm run lint (zero max-lines-per-function warnings in scope),
tsc --noEmit (0 errors), npm run build succeeds.
